requirements.txt
anthropic>=0.40.0
pypdf>=4.0.0
python-docx>=1.1.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0
